requirement.txt
Flask==2.2.5
Flask-WTF==1.1.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==2.3.7
